refactor: move specialized agents to theri own file
- Introduced `math-agent`, `research-agent`, and `weather-agent` with dedicated tools for calculations, information gathering, and weather data retrieval.
- Implemented corresponding functions and data structures for each agent to handle specific tasks.
- Updated README to include details about new agents and their functionalities.
